Refactor the project, put server side code to /server folder
