+++ package.json
@@ -2,8 +2,9 @@
   "name": "csgo-artnet",
   "version": "1.0.0",
   "description": "A small NodeJS application to transfer data from Counter-Strike: Global Offensive and translate it for a lighting console using the ArtNet protocol.",
-  "main": "index.js",
+  "main": "server/index.js",
   "scripts": {
+    "start": "node server/index.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
